TEC-4290: Renamed component name and its appropriate types and imports.
