--- src/components/NewPaymentMethods/PaymentMethodsContainer/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-import { NewPaymentMethodsContainer } from './style'
-import { NewPaymentMethodsProps } from '../type'
-
-const PaymentMethodsContainer: React.FC<NewPaymentMethodsProps> = ({
-  children,
-}) => {
-  return <NewPaymentMethodsContainer>{children}</NewPaymentMethodsContainer>
-}
-
-export { PaymentMethodsContainer }
