--- src/components/NewPaymentMethods/index.tsx
@@ -1,5 +1,0 @@
-export { PaymentMethodsContainer as Root } from './PaymentMethodsContainer'
-export { ContainteStyled as Content } from './PaymentMethodsComplement'
-export { PaymentMethodsTitle as Title } from './PaymentMethodsTitle'
-export { InstallmentStyle as Items } from './PaymentMethodsInstallment'
-export { PaymentMethodsSubItems as SubItems } from './PaymentMethodsSubItems'
