--- src/components/PaymentMethods/PaymentMethodsContainer/type.ts
@@ -1,11 +1,0 @@
-import React from 'react'
-
-type PaymentMethodsProps = {
-  children: React.ReactNode
-  icon?: React.ReactNode
-  label: string
-  labelColor?: string
-  installment?: string
-}
-
-export type { PaymentMethodsProps }
