--- src/components/PaymentMethods/PaymentMethodsLabel/type.ts
@@ -1,7 +1,0 @@
-type PaymentMethodsContainerProps = {
-  label: string
-  labelColor?: string
-  icon?: React.ReactNode
-}
-
-export type { PaymentMethodsContainerProps }
