--- src/components/PaymentMethods/PaymentsMethodsInstallment/type.ts
@@ -1,6 +1,0 @@
-type PaymentsMethodsItemProps = {
-  installment?: string
-  children: React.ReactNode
-}
-
-export type { PaymentsMethodsItemProps }
