+++ src/components/PaymentMethods/index.tsx
@@ -1,3 +1,5 @@
-export { PaymentMethods } from './PaymentMethodsContainer'
-export { PaymentMethodsLabel } from './PaymentMethodsLabel'
-export { PaymentContainerInstallment } from './PaymentsMethodsInstallment'
+export { PaymentMethodsContainer as Root } from './PaymentMethodsContainer'
+export { ContainteStyled as Content } from './PaymentMethodsComplement'
+export { PaymentMethodsTitle as Title } from './PaymentMethodsTitle'
+export { InstallmentStyle as Items } from './PaymentMethodsInstallment'
+export { PaymentMethodsSubItems as SubItems } from './PaymentMethodsSubItems'
