+++ src/components/PaymentMethods/type.ts
@@ -1,10 +1,10 @@
 import { HTMLAttributes } from 'react'
 
-type NewPaymentMethodsProps = {
+type PaymentMethodsProps = {
   label?: string | undefined
   icon?: JSX.Element
   installment?: string
   children?: React.ReactNode
 } & HTMLAttributes<HTMLDivElement>
 
-export type { NewPaymentMethodsProps }
+export type { PaymentMethodsProps }
